requirements.txt
pytube
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
numpy
matplotlib
ffmpeg-python